Add Puppet::Transaction::Report#userdata
Without this patch the transaction report created by the agent does not
have an attribute to store user defined data.  This is a problem because
we need to store at minimum the role classes the agent is enforcing to
make the data accessible via BQ and Data Studio via custom report
processor.

This patch addresses the problem by establishing an userdata map object.
The schema is defined by the end user.  The value must be serializable
to JSON encoding so that the data may be loaded into BQ.

Note that initialization of the monkey patching using app_defaults has
been selected because it currently lacks race conditions related to
Puppet initialization.  Ideally a future patch to puppet will establish
a supported hook in the application initialization process which will
find Gems and call an initializer method.  This callback mechanism would
avoid the need to override Application#app_defaults and provide a better
supported method to customize the behavior of Puppet.